% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectStore.get.r
\name{objectStore.get}
\alias{objectStore.get}
\title{Fuction for getting data from objectStore}
\usage{
objectStore.get(creds)
}
\arguments{
\item{creds}{your swift credentials, see README for further info}
}
\description{
This function pulls data from objectStore to your kernel
}
\examples{
#this example runs in DSX
creds <-list(auth_url = "https://identity.open.softlayer.com",project = "object_storage_92c67982_------",project_id = "7babac2c------------------",region = "dallas",user_id = "18aa------------------",domain_id = "2c------------------",domain_name =  "------",username = "admin_774cd5e------------------------",password = "i------------",container = "notebooks", filename = "yourfile.csv")
df <- objectStore.get(creds)
}
\keyword{swift}

