% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectStore.put.r
\name{objectStore.put}
\alias{objectStore.put}
\title{Fuction for saving data to ObjectStore}
\usage{
objectStore.put(creds, out_file)
}
\arguments{
\item{creds}{your swift credentials, see README for further info}
}
\description{
This function pushes data to objectStore from your kernel
}
\examples{
#this example runs in DSX
creds <-list(auth_url = "https://identity.open.softlayer.com",project = "object_storage_92c67982_------",project_id = "7babac2c------------------",region = "dallas",user_id = "18aa------------------",domain_id = "2c------------------",domain_name =  "------",username = "admin_774cd5e------------------------",password = "i------------",container = "notebooks")
write.csv(df,'myCSVforGit.csv')
status<- objectStore.put(creds,'myCSVforGit.csv')
status
}
\keyword{swift}

